yo_te_ciudo_API/requirements.txt
dj-database-url==0.5.0
Django==1.11.18
djangorestframework==3.9.1
Pillow==5.4.1
psycopg2==2.7.7
python-decouple==3.1
pytz==2018.9
